fix: settelment screen 스크롤 수정(#87)
